// src/app/models/package.ts
export interface Package {
    _id: any;
    package_id : any;
    active_delivery_id: any;
    description:any;
    created_user: any;
    modified_user: any;
    weight: any;
    width: any;
    height: any;
    depth: any;
    from_name:any;
    from_address:any;
    from_location : any;
    to_name:any;
    to_address:any;
    to_location : any;
    deliveries: any;
    from_location_lat: any,
    from_location_long: any,
    to_location_lat: any,
    to_location_long: any


}